requirements.txt
bs4==0.0.1
requests==2.25.1
python-slugify 5.0.2
